VectorPlotDigitizer

Documentation for VectorPlotDigitizer.

VectorPlotDigitizer.binedges_to_bin_midpoints

VectorPlotDigitizer.fill_zeros_bins

VectorPlotDigitizer.binedges_to_bin_midpoints — Method

binedges_to_bin_midpoints(data; zero_threshold::Real = 0.0, atol::Real= 0.0)

Transforms a set of points that delimits the vertexes of a binned plot into a set of points for each bin mid point value.

data must be two-column array of numbers that delimit some bin vertexes. The function will detect a bin when two consecutive values of the y axis are not vertical. In such case it will compute the midpoint value of the x axis and take that y axis value.

If a value of zero_threshold different than 0 is given, then every detected bin with a y value y < zero_threshold will be considered as 0.0.

If a value of atol is provided, this will be taken as the absolute tolerance in the isapprox method that checks if two points are a straigth vertical line or not. Higher values of atol may give undesired results.

VectorPlotDigitizer.fill_zeros_bins — Method

fill_zeros_bins(real_xs, xs, ys, atol; zero_threshold::Real = 0.0)

Given some digitized binned data, where xs are the the bins x axis midpoints and ys are the corresponding values on the y axis, this function fills with zeros the missing bins.

The complete set of bins is given by real_xs and atol is the absolute tolerance to use when comparing whether a given element of xs is inside of ys or not. Smaller values of atol will make less values of xs to be found inside real_xs.

If a value of zero_threshold different than 0 is given, then every bin with a y value y < zero_threshold will be considered as 0.0.

Examples

Given the following real_x bins and the digitized x_bins and y_bins:

julia> real_bins = [0.5, 1.5, 2.5, 3.5]
4-element Vector{Float64}:
 0.5
 1.5
 2.5
 3.5

julia> x_bins = [0.5001, 2.499987]
2-element Vector{Float64}:
 0.5001
 2.499987

julia> y_bins = [34.7, 4.2]
2-element Vector{Float64}:
 34.7
  4.2

We can fill it with zeros by choosing an absolute tolerance of atol = 1e-4:

julia> fill_zeros_bins(real_bins, x_binx, y_bins, 1e-4)
4-element Vector{Float64}:
 34.7
  0.0
  4.2
  0.0

We can also set a zero threshold at 10.0 so every point of y_bins smaller than that will bet zero:

julia> fill_zeros_bins(real_bins, x_binx, y_bins, 1e-4; zero_threshold = 10.0)
4-element Vector{Float64}:
 34.7
  0.0
  0.0
  0.0

Notes

A warning is given if the absolute tolerance is very small. This means that the function is filling with more zeros than needed and some ys values are not taken into account. This warning has nothing to do whether we set a high zero_threshold value or not.